+++ src/math/vector.rs
@@ -1,16 +1,18 @@
+use crate::Float;
+
 #[must_use]
-pub fn norm(x: &[f64]) -> f64 {
-    x.iter().map(|xi| xi * xi).sum::<f64>().sqrt()
+pub fn norm(x: &[Float]) -> Float {
+    x.iter().map(|xi| xi * xi).sum::<Float>().sqrt()
 }
 
 #[must_use]
 #[inline]
-pub fn dot(x: &[f64], y: &[f64]) -> f64 {
+pub fn dot(x: &[Float], y: &[Float]) -> Float {
     x.iter().zip(y).fold(0.0, |dot, (xi, yi)| dot + xi * yi)
 }
 
 #[must_use]
-pub fn cosine_similarity(x: &[f64], y: &[f64]) -> f64 {
+pub fn cosine_similarity(x: &[Float], y: &[Float]) -> Float {
     dot(x, y) / norm(x) / norm(y)
 }
 
@@ -23,7 +25,7 @@ mod tests {
         let vector_1 = [1.0, 2.0, 3.0];
         let vector_2 = [3.0, 5.0, 7.0];
         let similarity = cosine_similarity(&vector_1, &vector_2);
-        assert!((similarity - 0.9974149030430578).abs() < f64::EPSILON);
+        assert!((similarity - 0.9974149030430578).abs() < Float::EPSILON);
     }
 }
