+++ src/trainer.rs
@@ -18,6 +18,7 @@ use crate::trainer::loss::LossPair;
 use crate::trainer::math::make_gradient_descent_step;
 use crate::trainer::math::predict_probability;
 use crate::trainer::model::Model;
+use crate::Float;
 
 pub mod dataset;
 mod loss;
@@ -58,7 +59,7 @@ async fn run_epochs(
     opts: TrainerOpts,
     mut dataset: Dataset,
     should_stop: Arc<AtomicBool>,
-) -> crate::Result<f64> {
+) -> crate::Result<Float> {
     let mut last_losses = LossPair::infinity();
     let mut model = Model::new(redis, opts.model).await?;
 
@@ -84,9 +85,9 @@ async fn run_epochs(
                 losses.train,
                 losses.test,
                 model.regularization,
-                dataset.sample.len() as f64 / 1000.0 / start_instant.elapsed().as_secs_f64(),
-                dataset.sample.len() as f64 / 1000.0,
-                model.n_modified_accounts() as f64 / 1000.0,
+                dataset.sample.len() as Float / 1000.0 / start_instant.elapsed().as_secs_f64(),
+                dataset.sample.len() as Float / 1000.0,
+                model.n_modified_accounts() as Float / 1000.0,
                 model.n_initialized_accounts,
                 model.n_new_accounts,
             );
@@ -108,9 +109,9 @@ fn adjust_regularization(
     nr_epoch: usize,
     last_losses: &LossPair,
     losses: &LossPair,
-    regularization: f64,
-    auto_r_bump_chance: Option<f64>,
-) -> f64 {
+    regularization: Float,
+    auto_r_bump_chance: Option<Float>,
+) -> Float {
     if losses.test > last_losses.test {
         return if losses.train < last_losses.train {
             regularization + 0.001
@@ -144,7 +145,7 @@ async fn run_epoch(dataset: &mut Dataset, model: &mut Model) -> crate::Result<Lo
             .await?;
 
         let prediction = predict_probability(factors.vehicle, factors.account);
-        let label = point.n_wins as f64 / point.n_battles as f64;
+        let label = point.n_wins as Float / point.n_battles as Float;
 
         if !point.is_test {
             losses_builder.train.push_sample(prediction, label);
